% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AT.R
\name{AT}
\alias{AT}
\title{Title Adenine and Thymine nucleotide  basepair calculation}
\usage{
AT(x)
}
\arguments{
\item{x}{is the input sequence passed into  AT function  which contains character vector  "A","T","G","C"}
}
\value{
total percentage of Adenine and Thymine nucleoptide base pair in the input sequence
}
\description{
\code{AT} returns percentage of Adenine and Thymine nucleotide basepair composition in the input sequence
}
\examples{
test2<-AT("ATAGCATCAAGGTACGATGC")
test2
55
}
