% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdna.R
\name{cdna}
\alias{cdna}
\title{Title cdna synthesis}
\usage{
cdna(x)
}
\arguments{
\item{x}{is the input sequence passed into  cdna function  which contains character vector  "A","T","G","C"}
}
\value{
character vector ,complementary dna is synthesized by replacing "A" with "T" and "G" with "C"and viceversa
}
\description{
\code{cdna} returns the complementary dna for the input  sequence
}
