% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC.R
\name{GC}
\alias{GC}
\title{Title Guanine and Cytosine  nucleotide  basepair calculation}
\usage{
GC(x)
}
\arguments{
\item{x}{is the input sequence passed into  GC function  which contains character vector  "A","T","G","C"}
}
\description{
\code{GC} returns percentage of Guanine and Cytosine  nucleotide basepair composition in the input sequence
}
\examples{
test3<-GC("ATAGCATCAAGGTACGATGC")
test3
45
}
