% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingplot.R
\name{slidingplot}
\alias{slidingplot}
\title{Title sliding plot function}
\usage{
slidingplot(windowsize, inputseq)
}
\arguments{
\item{inputseq}{}
}
\value{
graphical / visual representation of input sequence
}
\description{
Title sliding plot function
}
\examples{
slidingplot(10,"AAAATTTGGGCCC")
}
